requirements.txt
Flask
pyTelegramBotAPI
twilio
mysql-connector-python
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client